move func to utils.common
